+++ packages/backend/src/index.ts
@@ -1,3 +1,3 @@
-export * from './router';
+export * from './request';
 export * from './rpc';
 export * from './db';
